third_party/WebKit/LayoutTests/animations/transition-and-animation-1.html

<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"
   "http://www.w3.org/TR/html4/loose.dtd">

<html lang="en">
<head>
  <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
  <title>Transition/Animation Test #1</title>
  <style type="text/css" media="screen">
    #box {
        position: absolute;
        left: 0;
        top: 100px;
        height: 100px;
        width: 100px;
        background-color: blue;
        animation-duration: 0.5s;
        animation-timing-function: linear;
        animation-name: anim;
        transition-property: transform;
        transition-duration: 10s;
    }
    @keyframes anim {
        from { transform: translateX(200px); }
        to   { transform: translateX(300px); }
    }
    </style>
    <script src="resources/animation-test-helpers.js" type="text/javascript" charset="utf-8"></script>
    <script type="text/javascript" charset="utf-8">
    
    const expectedValues = [
      // [time, element-id, property, expected-value, tolerance]
      [0.55, "box", "transform", "none", null],
    ];
    
    runAnimationTest(expectedValues, undefined, undefined, 'do-not-use-pause-api');
    
  </script>
</head>
<body>
This test has a transition and animation on the same property (transform). But the transition is never triggered,
so nothing should be moving when the animation finishes.
<div id="box">
</div>
<div id="result">
</div>
</body>
</html>
