third_party/WebKit/LayoutTests/platform/mac-lion/virtual/web-animations-css/animations/transition-and-animation-1-expected.txt

Warning this test is running in real-time and may be flaky.
FAIL - "webkitTransform" property for "box" element at 0.55s expected: none but saw: matrix(1, 0, 0, 1, 300, 0)

Log:
22.826000000009117 - runAnimationTest
25.36700000007386 - First webkitAnimationStart event fired
25.463000000058855 - Finished waiting for animations to start
25.559000000043852 - RAF callback, animation time: 0.06099999995967664
25.71899999998095 - RAF callback, animation time: 0.21700000002056186
26.349999999979445 - RAF callback, animation time: 0.8550000000013824
27.840999999966698 - RAF callback, animation time: 2.3529999999709617
30.828000000042266 - RAF callback, animation time: 5.34000000004653
34.474000000045635 - RAF callback, animation time: 8.987000000047374
36.85300000006464 - RAF callback, animation time: 11.365999999952692
37.90400000002592 - RAF callback, animation time: 12.417000000027656
39.76799999998093 - RAF callback, animation time: 14.277999999990243
40.92800000000807 - RAF callback, animation time: 15.442000000007283
42.146999999999935 - RAF callback, animation time: 16.660000000001673
43.327999999974054 - RAF callback, animation time: 17.836999999985892
44.392000000016196 - RAF callback, animation time: 18.894000000045708
47.196999999982836 - RAF callback, animation time: 21.705999999994674
48.33900000005542 - RAF callback, animation time: 22.853000000054635
49.626999999986765 - RAF callback, animation time: 24.139999999988504
50.80699999996341 - RAF callback, animation time: 25.318999999967673
52.19499999998334 - RAF callback, animation time: 26.708999999982552
54.3920000000071 - RAF callback, animation time: 28.874999999970896
55.4740000000038 - RAF callback, animation time: 29.989000000000487
56.58800000003339 - RAF callback, animation time: 31.10300000003008
57.74800000006053 - RAF callback, animation time: 32.26300000005722
58.898999999996704 - RAF callback, animation time: 33.41299999999592
60.04300000006424 - RAF callback, animation time: 34.55799999994724
61.38799999996536 - RAF callback, animation time: 35.90299999996205
62.54100000001017 - RAF callback, animation time: 37.045000000034634
63.99799999996958 - RAF callback, animation time: 38.50799999997889
65.15799999999672 - RAF callback, animation time: 39.67199999999593
66.41799999999876 - RAF callback, animation time: 40.931000000000495
67.89400000002388 - RAF callback, animation time: 42.40300000003572
69.02900000000045 - RAF callback, animation time: 43.54200000000219
70.10999999999967 - RAF callback, animation time: 44.62300000000141
71.26700000003439 - RAF callback, animation time: 45.7810000000336
72.37700000007408 - RAF callback, animation time: 46.89199999995708
73.6699999999928 - RAF callback, animation time: 48.18399999999201
74.85099999996692 - RAF callback, animation time: 49.36299999997118
76.01399999998648 - RAF callback, animation time: 50.527999999985695
77.17300000001615 - RAF callback, animation time: 51.68700000001536
78.34200000002056 - RAF callback, animation time: 52.856000000019776
79.52099999999973 - RAF callback, animation time: 54.03599999999642
81.53700000002573 - RAF callback, animation time: 56.048000000032516
82.69400000006044 - RAF callback, animation time: 57.20799999994597
83.84799999998904 - RAF callback, animation time: 58.36199999998826
85.90900000001511 - RAF callback, animation time: 60.4200000000219
88.02400000001853 - RAF callback, animation time: 62.53500000002532
89.56399999999576 - RAF callback, animation time: 64.0769999999975
92.30200000001787 - RAF callback, animation time: 66.81300000002466
93.69300000003022 - RAF callback, animation time: 68.20000000004711
94.91900000000442 - RAF callback, animation time: 69.43100000000868
96.08000000002903 - RAF callback, animation time: 70.5920000000333
97.30500000000575 - RAF callback, animation time: 71.81900000000496
99.47399999998652 - RAF callback, animation time: 73.9849999999933
101.67100000001028 - RAF callback, animation time: 76.18200000001707
103.46800000002077 - RAF callback, animation time: 77.97900000002755
104.65299999998479 - RAF callback, animation time: 79.166999999984
106.13599999999224 - RAF callback, animation time: 80.64899999999398
107.29800000001433 - RAF callback, animation time: 81.81200000001354
108.64700000001903 - RAF callback, animation time: 83.16100000001825
109.81300000003102 - RAF callback, animation time: 84.32800000002771
110.96700000007331 - RAF callback, animation time: 85.48299999995379
112.14000000006763 - RAF callback, animation time: 86.65499999995063
113.33400000000893 - RAF callback, animation time: 87.84800000000814
114.49000000004617 - RAF callback, animation time: 89.00500000004286
115.97100000005867 - RAF callback, animation time: 90.48500000005788
117.21000000000004 - RAF callback, animation time: 91.72499999999673
118.3770000000095 - RAF callback, animation time: 92.89200000000619
119.93100000006507 - RAF callback, animation time: 94.44599999994807
122.1180000000004 - RAF callback, animation time: 96.62600000001476
124.04000000003634 - RAF callback, animation time: 98.54600000005576
124.98200000004545 - RAF callback, animation time: 99.49600000004466
126.03500000000167 - RAF callback, animation time: 100.54700000000594
127.25899999998092 - RAF callback, animation time: 101.77099999998518
128.81800000002386 - RAF callback, animation time: 103.32900000003065
130.46700000006695 - RAF callback, animation time: 104.97799999996005
131.86900000005153 - RAF callback, animation time: 106.38100000005579
135.8220000000756 - RAF callback, animation time: 110.33499999996366
137.14900000002217 - RAF callback, animation time: 111.66300000002138
139.4669999999678 - RAF callback, animation time: 113.97899999997207
140.62000000001262 - RAF callback, animation time: 115.13400000001184
142.25999999996475 - RAF callback, animation time: 116.77399999996396
143.46199999999953 - RAF callback, animation time: 117.97599999999875
144.83400000005986 - RAF callback, animation time: 119.34799999994539
146.00300000006428 - RAF callback, animation time: 120.51799999994728
147.1920000000182 - RAF callback, animation time: 121.70600000001741
148.6170000000584 - RAF callback, animation time: 123.13100000005761
149.90199999999732 - RAF callback, animation time: 124.41200000000663
154.51200000006793 - RAF callback, animation time: 129.02099999996608
156.8520000000717 - RAF callback, animation time: 131.36199999996734
158.01799999997002 - RAF callback, animation time: 132.53099999997175
159.17699999999968 - RAF callback, animation time: 133.69000000000142
160.40299999997387 - RAF callback, animation time: 134.91799999997056
161.58000000007178 - RAF callback, animation time: 136.0939999999573
162.73899999998775 - RAF callback, animation time: 137.25399999998444
163.90200000000732 - RAF callback, animation time: 138.417000000004
165.63200000007328 - RAF callback, animation time: 140.14499999996133
166.73200000002453 - RAF callback, animation time: 141.24600000002374
168.09699999998884 - RAF callback, animation time: 142.43499999997766
174.5379999999841 - RAF callback, animation time: 148.95000000001346
175.87700000001405 - RAF callback, animation time: 150.3770000000486
177.1929999999884 - RAF callback, animation time: 151.7069999999876
178.45699999998033 - RAF callback, animation time: 152.96699999998964
179.61600000001 - RAF callback, animation time: 154.12900000001173
181.0689999999795 - RAF callback, animation time: 155.5829999999787
184.38000000003285 - RAF callback, animation time: 158.89000000004216
185.90600000004542 - RAF callback, animation time: 160.41900000004716
187.12300000004234 - RAF callback, animation time: 161.6350000000466
188.34100000003673 - RAF callback, animation time: 162.85500000003594
189.77000000006683 - RAF callback, animation time: 164.28399999995236
190.92799999998533 - RAF callback, animation time: 165.44299999998202
192.08100000003014 - RAF callback, animation time: 166.59600000002683
193.224999999984 - RAF callback, animation time: 167.73999999998068
194.38500000001113 - RAF callback, animation time: 168.90000000000782
197.133000000008 - RAF callback, animation time: 171.64400000001478
199.0369999999757 - RAF callback, animation time: 173.54899999997997
201.24799999996412 - RAF callback, animation time: 175.7599999999684
202.42700000005698 - RAF callback, animation time: 176.9410000000562
203.61000000002605 - RAF callback, animation time: 178.12400000002526
204.72800000004554 - RAF callback, animation time: 179.24300000004223
205.8640000000196 - RAF callback, animation time: 180.37900000001628
206.93100000005415 - RAF callback, animation time: 181.44500000005337
208.08099999999286 - RAF callback, animation time: 182.59599999998954
209.30699999996705 - RAF callback, animation time: 183.82099999996626
210.78499999998712 - RAF callback, animation time: 185.2999999999838
211.97700000004716 - RAF callback, animation time: 186.49200000004384
213.1370000000743 - RAF callback, animation time: 187.6519999999573
214.33999999999287 - RAF callback, animation time: 188.85499999998956
215.72700000001532 - RAF callback, animation time: 190.24100000001454
216.90300000000207 - RAF callback, animation time: 191.41700000000128
218.06100000003426 - RAF callback, animation time: 192.57600000003094
219.21399999996538 - RAF callback, animation time: 193.72999999995955
220.36800000000767 - RAF callback, animation time: 194.8849999999993
221.51300000007268 - RAF callback, animation time: 196.02899999995316
222.67499999998108 - RAF callback, animation time: 197.19099999997525
223.84499999998297 - RAF callback, animation time: 198.35999999997966
224.97499999997217 - RAF callback, animation time: 199.4919999999638
226.12900000001446 - RAF callback, animation time: 200.64500000000862
227.27199999997083 - RAF callback, animation time: 201.78699999996752
228.411000000051 - RAF callback, animation time: 202.9250000000502
229.57700000006298 - RAF callback, animation time: 204.09199999994598
230.70400000005975 - RAF callback, animation time: 205.21900000005644
231.86499999997068 - RAF callback, animation time: 206.3789999999699
232.96100000004571 - RAF callback, animation time: 207.4760000000424
234.1670000000704 - RAF callback, animation time: 208.6819999999534
235.4050000000143 - RAF callback, animation time: 209.91300000002866
236.63199999998596 - RAF callback, animation time: 211.14699999998265
237.79500000000553 - RAF callback, animation time: 212.30900000000474
238.94000000007054 - RAF callback, animation time: 213.45499999995354
240.0800000000345 - RAF callback, animation time: 214.59600000002865
241.21000000002368 - RAF callback, animation time: 215.72500000002037
242.3310000000356 - RAF callback, animation time: 216.8460000000323
243.45200000004752 - RAF callback, animation time: 217.9670000000442
245.08100000002742 - RAF callback, animation time: 219.5960000000241
246.27699999996366 - RAF callback, animation time: 220.79099999996288
247.42800000001353 - RAF callback, animation time: 221.9440000000077
248.61599999996997 - RAF callback, animation time: 223.1299999999692
249.75300000005518 - RAF callback, animation time: 224.26800000005187
250.88000000005195 - RAF callback, animation time: 225.39500000004864
251.99700000007397 - RAF callback, animation time: 226.51199999995697
253.11299999998482 - RAF callback, animation time: 227.6279999999815
254.2160000000422 - RAF callback, animation time: 228.73100000003888
255.33800000005158 - RAF callback, animation time: 229.85300000004827
256.52500000001055 - RAF callback, animation time: 231.01599999995415
257.6599999999871 - RAF callback, animation time: 232.17399999998634
258.8230000000067 - RAF callback, animation time: 233.33800000000338
259.9680000000717 - RAF callback, animation time: 234.48399999995218
261.10800000003564 - RAF callback, animation time: 235.6240000000298
262.26500000007036 - RAF callback, animation time: 236.77999999995336
263.41500000000906 - RAF callback, animation time: 237.92900000000827
264.60500000007414 - RAF callback, animation time: 239.11499999996977
265.7320000000709 - RAF callback, animation time: 240.24599999995644
266.8740000000298 - RAF callback, animation time: 241.38700000003155
268.0410000000393 - RAF callback, animation time: 242.5550000000385
269.15600000006634 - RAF callback, animation time: 243.67099999994934
270.24900000003527 - RAF callback, animation time: 244.76300000003448
271.4300000000094 - RAF callback, animation time: 245.9440000000086
272.59700000001885 - RAF callback, animation time: 247.11100000001807
273.77100000001064 - RAF callback, animation time: 248.28500000000986
275.01600000005055 - RAF callback, animation time: 249.53000000004977
276.2050000000045 - RAF callback, animation time: 250.7190000000037
308.65800000003674 - RAF callback, animation time: 283.16800000004605
309.80799999997544 - RAF callback, animation time: 284.3209999999772
310.9540000000379 - RAF callback, animation time: 285.46700000003966
312.09499999999935 - RAF callback, animation time: 286.6080000000011
313.3189999999786 - RAF callback, animation time: 287.8319999999803
314.46600000003855 - RAF callback, animation time: 288.98000000003776
316.50200000001405 - RAF callback, animation time: 290.997999999945
317.8500000000213 - RAF callback, animation time: 292.3640000000205
319.16400000000067 - RAF callback, animation time: 293.6770000000024
320.43099999998503 - RAF callback, animation time: 294.94499999998425
321.6180000000577 - RAF callback, animation time: 296.11799999997857
322.74500000005446 - RAF callback, animation time: 297.2590000000537
324.1010000000415 - RAF callback, animation time: 298.6150000000407
325.26900000004844 - RAF callback, animation time: 299.78300000004765
326.4330000000655 - RAF callback, animation time: 300.9479999999485
327.5849999999991 - RAF callback, animation time: 302.1009999999933
328.7119999999959 - RAF callback, animation time: 303.2269999999926
329.88799999998264 - RAF callback, animation time: 304.40199999998185
331.07700000005025 - RAF callback, animation time: 305.59100000004946
332.6270000000022 - RAF callback, animation time: 307.14000000000397
333.9780000000019 - RAF callback, animation time: 308.4929999999986
335.1559999999836 - RAF callback, animation time: 309.6699999999828
336.20700000005854 - RAF callback, animation time: 310.72100000005776
337.71600000000035 - RAF callback, animation time: 312.22999999999956
338.88500000000477 - RAF callback, animation time: 313.39700000000903
340.0760000000673 - RAF callback, animation time: 314.5849999999655
341.4539999999988 - RAF callback, animation time: 315.9650000000056
342.583999999988 - RAF callback, animation time: 317.0979999999872
343.7579999999798 - RAF callback, animation time: 318.271999999979
345.2190000000428 - RAF callback, animation time: 319.73200000004454
346.49100000001454 - RAF callback, animation time: 321.00500000001375
347.725999999966 - RAF callback, animation time: 322.23899999996775
348.88599999999315 - RAF callback, animation time: 323.39999999999236
350.03500000004806 - RAF callback, animation time: 324.55000000004475
351.16500000003725 - RAF callback, animation time: 325.6810000000314
352.2960000000239 - RAF callback, animation time: 326.8110000000206
353.4230000000207 - RAF callback, animation time: 327.9380000000174
354.52199999997447 - RAF callback, animation time: 329.03699999997116
355.6310000000167 - RAF callback, animation time: 330.1460000000134
356.8760000000566 - RAF callback, animation time: 331.3900000000558
357.97600000000784 - RAF callback, animation time: 332.49000000000706
359.1030000000046 - RAF callback, animation time: 333.6180000000013
360.2130000000443 - RAF callback, animation time: 334.72900000003847
361.3080000000082 - RAF callback, animation time: 335.82400000000234
362.41500000005544 - RAF callback, animation time: 336.9310000000496
363.5980000000245 - RAF callback, animation time: 338.1130000000212
364.77100000001883 - RAF callback, animation time: 339.2860000000155
365.93500000003587 - RAF callback, animation time: 340.45000000003256
367.13299999996707 - RAF callback, animation time: 341.6469999999663
368.30700000007255 - RAF callback, animation time: 342.822999999953
369.4719999999734 - RAF callback, animation time: 343.98699999997007
370.63399999999547 - RAF callback, animation time: 345.14999999998963
371.8260000000555 - RAF callback, animation time: 346.3410000000522
372.9809999999816 - RAF callback, animation time: 347.49699999997574
374.23999999998614 - RAF callback, animation time: 348.75499999998283
375.42000000007647 - RAF callback, animation time: 349.9349999999595
376.5680000000202 - RAF callback, animation time: 351.0840000000144
377.7750000000424 - RAF callback, animation time: 352.2880000000441
378.88699999996334 - RAF callback, animation time: 353.4029999999575
380.0469999999905 - RAF callback, animation time: 354.5599999999922
381.18199999996705 - RAF callback, animation time: 355.69599999996626
382.310000000075 - RAF callback, animation time: 356.824999999958
383.4360000000743 - RAF callback, animation time: 357.9509999999573
384.537000000023 - RAF callback, animation time: 359.0510000000222
385.70900000001984 - RAF callback, animation time: 360.22300000001906
386.8370000000141 - RAF callback, animation time: 361.3520000000108
388.00000000003365 - RAF callback, animation time: 362.51500000003034
389.1720000000305 - RAF callback, animation time: 363.68800000002466
390.34300000002986 - RAF callback, animation time: 364.859000000024
391.5060000000494 - RAF callback, animation time: 366.0220000000436
392.68300000003364 - RAF callback, animation time: 367.19800000003033
393.8510000000406 - RAF callback, animation time: 368.3660000000373
395.07400000002235 - RAF callback, animation time: 369.5870000000241
396.2579999999889 - RAF callback, animation time: 370.77099999999064
397.4190000000135 - RAF callback, animation time: 371.9340000000102
398.5800000000381 - RAF callback, animation time: 373.09400000003734
400.0740000000178 - RAF callback, animation time: 374.5890000000145
401.2490000000071 - RAF callback, animation time: 375.76400000000376
402.36800000002404 - RAF callback, animation time: 376.8840000000182
403.60599999996793 - RAF callback, animation time: 378.11999999996715
404.7130000000152 - RAF callback, animation time: 379.2280000000119
405.8630000000676 - RAF callback, animation time: 380.3779999999506
407.0330000000695 - RAF callback, animation time: 381.54899999994996
408.1499999999778 - RAF callback, animation time: 382.665999999972
409.3110000000024 - RAF callback, animation time: 383.8269999999966
410.4899999999816 - RAF callback, animation time: 385.0049999999783
411.6760000000568 - RAF callback, animation time: 386.19100000005346
412.8680000000031 - RAF callback, animation time: 387.3839999999973
414.0590000000657 - RAF callback, animation time: 388.57499999994616
415.16300000000683 - RAF callback, animation time: 389.6780000000035
416.3090000000693 - RAF callback, animation time: 390.8239999999523
417.51299999998537 - RAF callback, animation time: 392.0269999999846
418.66700000002766 - RAF callback, animation time: 393.18200000002435
419.8230000000649 - RAF callback, animation time: 394.3389999999454
420.9859999999708 - RAF callback, animation time: 395.50199999996494
422.1900000000005 - RAF callback, animation time: 396.7000000000098
423.3649999999898 - RAF callback, animation time: 397.8819999999814
424.5660000000271 - RAF callback, animation time: 399.0810000000238
425.73700000002646 - RAF callback, animation time: 400.2540000000181
426.88299999997525 - RAF callback, animation time: 401.39799999997194
428.0630000000656 - RAF callback, animation time: 402.5779999999486
429.22699999996894 - RAF callback, animation time: 403.7429999999631
430.4260000000113 - RAF callback, animation time: 404.9370000000181
431.5750000000662 - RAF callback, animation time: 406.08899999995174
434.7340000000486 - RAF callback, animation time: 409.24600000005285
441.5950000000066 - RAF callback, animation time: 416.10500000001593
442.84800000002633 - RAF callback, animation time: 417.363000000023
444.0270000000055 - RAF callback, animation time: 418.5420000000022
445.1490000000149 - RAF callback, animation time: 419.6630000000141
446.2110000000621 - RAF callback, animation time: 420.7249999999476
447.3830000000589 - RAF callback, animation time: 421.89399999995203
448.5329999999976 - RAF callback, animation time: 423.0479999999943
449.6990000000096 - RAF callback, animation time: 424.21300000000883
450.8530000000519 - RAF callback, animation time: 425.3680000000486
452.0899999999983 - RAF callback, animation time: 426.6020000000026
453.2239999999774 - RAF callback, animation time: 427.7389999999741
454.3459999999868 - RAF callback, animation time: 428.8609999999835
455.4440000000568 - RAF callback, animation time: 429.9590000000535
456.6130000000612 - RAF callback, animation time: 431.1280000000579
457.74200000005294 - RAF callback, animation time: 432.2550000000547
458.8730000000396 - RAF callback, animation time: 433.38600000004135
460.02599999997074 - RAF callback, animation time: 434.5409999999674
461.1810000000105 - RAF callback, animation time: 435.69400000001224
462.3380000000452 - RAF callback, animation time: 436.8520000000444
463.5050000000547 - RAF callback, animation time: 438.0190000000539
465.51399999998466 - RAF callback, animation time: 440.0269999999864
466.85600000000704 - RAF callback, animation time: 441.3690000000088
467.98499999999876 - RAF callback, animation time: 442.4980000000005
469.3600000000515 - RAF callback, animation time: 443.8740000000507
470.9139999999934 - RAF callback, animation time: 445.4279999999926
472.323000000074 - RAF callback, animation time: 446.8369999999595
473.50600000004306 - RAF callback, animation time: 448.0200000000423
475.19999999997253 - RAF callback, animation time: 449.71399999997175
476.6980000000558 - RAF callback, animation time: 451.2099999999464
478.0759999999873 - RAF callback, animation time: 452.588999999989
479.3160000000398 - RAF callback, animation time: 453.83000000003904
480.6550000000698 - RAF callback, animation time: 455.1689999999553
481.9840000000113 - RAF callback, animation time: 456.499000000008
483.1860000000461 - RAF callback, animation time: 457.7000000000453
484.3630000000303 - RAF callback, animation time: 458.8770000000295
485.5330000000322 - RAF callback, animation time: 460.0480000000289
486.6819999999734 - RAF callback, animation time: 461.1969999999701
487.824000000046 - RAF callback, animation time: 462.3390000000427
488.962000000015 - RAF callback, animation time: 463.4760000000142
490.0910000000067 - RAF callback, animation time: 464.60500000000593
491.6600000000244 - RAF callback, animation time: 466.1750000000211
492.8549999999632 - RAF callback, animation time: 467.3689999999624
493.99700000003577 - RAF callback, animation time: 468.511000000035
495.5730000000358 - RAF callback, animation time: 470.0880000000325
496.93999999999505 - RAF callback, animation time: 471.45399999999427
498.2710000000452 - RAF callback, animation time: 472.7850000000444
499.63900000000194 - RAF callback, animation time: 474.1520000000037
502.08600000007664 - RAF callback, animation time: 476.5989999999647
504.0089999999964 - RAF callback, animation time: 478.5219999999981
505.06300000006377 - RAF callback, animation time: 479.57499999995434
508.07399999996505 - RAF callback, animation time: 482.58499999997184
509.6019999999726 - RAF callback, animation time: 484.1149999999743
510.79200000003766 - RAF callback, animation time: 485.3050000000394
512.0040000000472 - RAF callback, animation time: 486.51600000005146
513.4010000000444 - RAF callback, animation time: 487.91400000004614
515.184999999974 - RAF callback, animation time: 489.69899999997324
516.8220000000474 - RAF callback, animation time: 491.3340000000517
518.0470000000241 - RAF callback, animation time: 492.56000000002587
519.506999999976 - RAF callback, animation time: 494.0199999999777
520.968000000039 - RAF callback, animation time: 495.4800000000432
522.2109999999702 - RAF callback, animation time: 496.7229999999745
523.4350000000632 - RAF callback, animation time: 497.94999999994616
524.6329999999944 - RAF callback, animation time: 499.14799999999104
525.9909999999763 - RAF callback, animation time: 500.5039999999781
527.2620000000643 - RAF callback, animation time: 501.77499999995234
528.5750000000462 - RAF callback, animation time: 503.0890000000454
529.749000000038 - RAF callback, animation time: 504.2630000000372
530.982999999992 - RAF callback, animation time: 505.4959999999937
532.2039999999788 - RAF callback, animation time: 506.717999999978
533.6970000000747 - RAF callback, animation time: 508.2109999999602
535.0260000000162 - RAF callback, animation time: 509.5410000000129
536.1890000000358 - RAF callback, animation time: 510.70400000003247
538.5929999999917 - RAF callback, animation time: 513.1069999999909
539.7830000000567 - RAF callback, animation time: 514.297000000056
540.9280000000081 - RAF callback, animation time: 515.4440000000022
542.2520000000759 - RAF callback, animation time: 516.7659999999614
544.4549999999708 - RAF callback, animation time: 518.96899999997
545.9540000000516 - RAF callback, animation time: 520.4680000000508
548.2580000000326 - RAF callback, animation time: 522.7720000000318
549.443999999994 - RAF callback, animation time: 523.9579999999933
550.8730000000241 - RAF callback, animation time: 525.3860000000259
552.1459999999934 - RAF callback, animation time: 526.6599999999926
553.3550000000105 - RAF callback, animation time: 527.8680000000122
554.7179999999798 - RAF callback, animation time: 529.231999999979
555.9440000000677 - RAF callback, animation time: 530.4539999999633
557.1780000000217 - RAF callback, animation time: 531.6870000000335
558.3740000000716 - RAF callback, animation time: 532.8849999999647
559.6020000000408 - RAF callback, animation time: 534.1130000000476
560.8429999999771 - RAF callback, animation time: 535.3509999999915
562.546999999995 - RAF callback, animation time: 537.0570000000043
563.6460000000625 - RAF callback, animation time: 538.1609999999455
564.8270000000366 - RAF callback, animation time: 539.3420000000333
566.0120000000006 - RAF callback, animation time: 540.5269999999973
567.2120000000405 - RAF callback, animation time: 541.7270000000372
568.4039999999868 - RAF callback, animation time: 542.9189999999835
569.6010000000342 - RAF callback, animation time: 544.1160000000309
570.7949999999755 - RAF callback, animation time: 545.3099999999722
571.9910000000255 - RAF callback, animation time: 546.5060000000221
573.0970000000752 - RAF callback, animation time: 547.6139999999532
574.3079999999736 - RAF callback, animation time: 548.8209999999754
575.4910000000564 - RAF callback, animation time: 550.0050000000556
575.533000000064 - Running checks for time: 550, delay: 0.005000000055588316
575.6979999999885 - Checking expectation: [null,0.55,"box","webkitTransform","none",null]
576.9530000000032 - Ending test
